Create Floyds tortoise and hare algorithm .py
This algorithm is called the 🐢 and 🐰 algorithm or Floyd’s cycle-finding algorithm after Robert Floyd, who invented it in the 1960s. The algorithm itself isn’t too complicated, but its explanation is pretty confusing.
The algorithm starts two objects called tortoise and hare moving through the list at different speeds starting at the beginning of the list. The tortoise moves one cell per step. The hare moves two cells per step.
If the hare reaches a link that is null, then the list has an end, so there is no loop.
If the list does contain a loop, the hare eventually enters the loop and starts running laps around it.
Meanwhile, the tortoise plods along until it eventually reaches the loop, too. At that point, both the tortoise and hare are inside the loop.